requirements.txt
aiosignal
aiosqlite>=0.19.0
aiohttp
asqlite
astunparse>=1.6.3
async-timeout
attrs
beautifulsoup4
certifi>=2022.6.15
cffi>=1.15.1
charset-normalizer
click>=8.1.3
colorama
cycler>=0.11.0
dateparser
decorator>=5.1.1
deep-translator
discord.py
emoji
flatbuffers>=22.12.6
fonttools
frozenlist
gast>=0.4.0
h5py
idna>=3.7
ImageHash>=4.3.1
ipykernel>=6.22.0
ipython>=8.12.0
ipywidgets
jarowinkler
Jinja2>=3.1.2
joblib>=1.1.0
keras
kiwisolver>=1.4.4
langdetect-py
Levenshtein
lingua-language-detector
matplotlib
multidict
nltk
numpy
oauthlib>=3.2.2
openai
packaging
pandas
Pillow
psutil>=5.9.4
pyasn1>=0.4.8
pyparsing
python-dateutil
python-dotenv
python-dotenv
pytz
PyYAML>=6.0
rapidfuzz>=2.3.0
regex
requests
rsa>=4.9
scikit-learn
scikit-learn
scipy>=1.9.1
setuptools
six>=1.16.0
soupsieve>=2.3.2.post1
tensorboard>=2.11.0
termcolor>=2.1.1
threadpoolctl>=3.1.0
tornado>=6.2
tqdm>=4.64.0
urllib3>=1.26.10
vaderSentiment
Werkzeug>=2.2.2
wrapt>=1.14.1
yarl